+++ base/system/sys_info_nacl.cc
@@ -11,7 +11,7 @@ namespace base {
 
 // static
 size_t SysInfo::VMAllocationGranularity() {
-  return getpagesize();
+  return static_cast<size_t>(getpagesize());
 }
 
 }  // namespace base
